Create TrellowTracker.g
Still super new. I wrote this up to track two of my employees in Trello so there may be code to calculate very specific items on my Trello Board. I'd like to expand this for general use eventually.

Most of this starter code is based on what I found here: http://kevinpelgrims.com/blog/2012/03/06/project-progress-tracking-with-google-docs-and-trello